requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.6
httpx==0.25.2
langextract==1.0.9
python-multipart==0.0.6
langdetect==1.0.9
cachetools==5.3.2
tenacity==8.2.3
pytesseract==0.3.10
pillow==10.1.0
prometheus-client==0.19.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
python-dotenv==1.0.0